KLIFS: Move back KLIFS_CLIENT/KLIFS_API_DEFINITIONS from utils to remote
