+++ opencadd/databases/klifs/remote.py
@@ -26,6 +26,9 @@ from opencadd.io import DataFrame, Rdkit
 
 _logger = logging.getLogger(__name__)
 
+KLIFS_API_DEFINITIONS = "https://dev.klifs.net/swagger_v2/swagger.json"
+KLIFS_CLIENT = SwaggerClient.from_url(KLIFS_API_DEFINITIONS, config={"validate_responses": False})
+
 
 class SerializableSwaggerClient(SwaggerClient):
     # Since they are using __attributes to mangle the namespace
