+++ opencadd/databases/klifs/utils.py
@@ -13,9 +13,6 @@ from bravado.client import SwaggerClient
 
 _logger = logging.getLogger(__name__)
 
-KLIFS_API_DEFINITIONS = "https://dev.klifs.net/swagger_v2/swagger.json"
-KLIFS_CLIENT = SwaggerClient.from_url(KLIFS_API_DEFINITIONS, config={"validate_responses": False})
-
 PATH_DATA = Path(__file__).parent / ".." / ".." / "data"
